// client/src/app/forms/interfaces/enums.ts
export enum FormComponentType {
  PARAGRAPH = 'paragraph',
  TEXT_INPUT = 'text_input',
  SINGLE_SELECT = 'single_select',
  MULTI_SELECT = 'multi_select',
  DATETIME = 'datetime',
  FILE = 'file',
  LOCATION = 'location',
}

export const enum DateFormat {
  DATE = 'date',
  DATETIME = 'datetime',
  TIME = 'time',
}

export const enum FileType {
  DOCUMENT = 'document',
  PDF = 'pdf',
  VIDEO = 'video',
  IMAGE = 'image',
  AUDIO = 'audio',
}

export const enum KeyboardType {
  DEFAULT = 'default',
  AUTO_CAPITALIZED = 'auto_capitalized',
  EMAIL = 'email',
  URL = 'url',
  PHONE = 'phone',
  NUMBER = 'number',
  DECIMAL = 'decimal',
  PASSWORD = 'password',
  NUMBER_PASSWORD = 'number_password',
}

export const enum OptionType {
  SHOW_DESCRIPTION,
  ENABLE_VALIDATION,
  GOTO_SECTION_BASED_ON_ANSWER,
  DELETE_ALL_RESPONSES,
  SENSITIVE_INFORMATION,
}